requirements.txt
PyGithub==2.8.1
PyYAML==6.0.2
tiktoken==0.11.0
tabulate==0.9.0
